Finalize already verifies the signature
The overview suggests that clients have to verify the signature
returned by the finalization function.

Clarify that this is not necessary, as it is already done by the
finalization function before returning.

This is obvious when directly implementing the protocol, but it
may be way less obvious when using higher-level APIs.