Try formatting the algorithms differently
In HTML, the current setup looks poor.  This formats the algorithms
using HTML.

I've only done this for Blind(), but it should probably be done for all
of them.

One note here in case you don't like this.  Step 1 mentioned
"sLenInBytes", but the variable was "sLen".  I fixed that here, but you
probably don't want to lose that change.